scripts/chapter4/pyproject.toml
[project]
name = "chapter4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain==1.0.3",
    "langchain-openai==1.0.1",
    "langchain-community==0.4.1",
    "langgraph==1.0.2",
    "python-dotenv==1.2.1",
    "jupyterlab==4.4.10",
]
